+++ server/src/completion/SnippetsCompletionProvider.ts
